requirements.txt
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mysql-connector==2.1.4
Werkzeug==0.12.2
